move unit tests into separate directory and modified tox.ini
