+++ test/unit/objects/domain_test.py
@@ -1,4 +1,4 @@
-from test.base import ClientBaseCase
+from test.unit.base import ClientBaseCase
 
 from linode_api4.objects import Domain, DomainRecord
