+++ test/unit/objects/mapped_object_test.py
