+++ test/unit/objects/polling_test.py
