+++ tox.ini
@@ -11,6 +11,6 @@ deps =
     httpretty
 commands =
     python setup.py install
-    coverage run --source linode_api4 -m pytest test/objects
+    coverage run --source linode_api4 -m pytest test/unit
     coverage report
     pylint linode_api4
